General feel.
