+++ app/src/main/java/ru/fabulus/fabulustrade/mvp/presenter/MainPresenter.kt
@@ -53,12 +53,8 @@ class MainPresenter : MvpPresenter<MainView>() {
         router.replaceScreen(Screens.tradersMainScreen(null))
     }
 
-    fun generalFeedClicked(){
-        if(profile.user != null){
-            router.replaceScreen(Screens.generalFeedFragment())
-        }else{
-            router.replaceScreen(Screens.signInScreen(false))
-        }
+    fun generalFeedClicked() {
+        router.replaceScreen(Screens.generalFeedFragment())
     }
 
     fun observationMenuClicked() {
