+++ app/src/main/java/ru/fabulus/fabulustrade/mvp/view/generalfeed/GeneralFeedPostView.kt
@@ -1,0 +1,10 @@
+package ru.fabulus.fabulustrade.mvp.view.generalfeed
+
+import moxy.MvpView
+import moxy.viewstate.strategy.AddToEndStrategy
+import moxy.viewstate.strategy.StateStrategyType
+
+@StateStrategyType(AddToEndStrategy::class)
+interface GeneralFeedPostView : MvpView {
+    fun init()
+}
