+++ app/src/main/java/ru/fabulus/fabulustrade/navigation/Screens.kt
@@ -4,6 +4,7 @@ import com.github.terrakok.cicerone.androidx.FragmentScreen
 import ru.fabulus.fabulustrade.mvp.model.entity.*
 import ru.fabulus.fabulustrade.ui.fragment.*
 import ru.fabulus.fabulustrade.ui.fragment.entrance.*
+import ru.fabulus.fabulustrade.ui.fragment.generalfeed.GeneralFeedFragment
 import ru.fabulus.fabulustrade.ui.fragment.subscriber.SubscriberMainFragment
 import ru.fabulus.fabulustrade.ui.fragment.subscriber.SubscriberNewsFragment
 import ru.fabulus.fabulustrade.ui.fragment.subscriber.SubscriberObservationFragment
