+++ app/src/main/res/layout/fragment_general_feed.xml
@@ -3,7 +3,11 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".ui.fragment.GeneralFeedFragment">
+    tools:context=".ui.fragment.generalfeed.GeneralFeedFragment">
 
+    <androidx.recyclerview.widget.RecyclerView
+        android:id="@+id/rv_general_feed"
+        android:layout_width="match_parent"
+        android:layout_height="match_parent" />
 
 </FrameLayout>
